requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database & async drivers
psycopg2-binary==2.9.10
asyncpg==0.29.0

# Message Queue
aiokafka==0.10.0

# Cache
redis[asyncio]==5.0.4

# Data processing & ML libraries
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.3.2

# Deep Learning
torch==2.1.1
transformers==4.41.0
onnxruntime==1.18.0

# Gradient Boosting
xgboost==2.0.3

# Model interpretability
shap==0.45.0

# HTTP clients
httpx==0.28.1
aiohttp==3.9.0
httpcore==1.0.9

# Data validation
pydantic==2.7.1

# Testing & Development
pytest==9.0.1
locust==2.43.4
